Feat: updated insertAssetData to ensure id does not skip numbers.
If a row with the highest current id would be deleted in std_assets,
it's number would be retained in the database. This retained number
would have 1 added to it and be set as the next inserted rows
information, regardless of whether the numbers flowed correctly in the
database.

I.e., if the database currently had 1, 2, 3, 4 as the info stored in the
id column, if 2, 3, and 4 were deleted, then the next row added would
still be listed as having an id of 5.