Feat: add functionality to retrieve attributes for a given type.
Implemented a new endpoint to dynamically fetch attributes associated
with a selected type from the database. The endpoint
'/attributes/{type}' returns a list of attributes based on the provided
type. This functionality enhances the application's ability to
dynamically populate attribute fields in the UI based on the selected
asset type.