Feat: adapted methods to get data for type fields.
Updated the get mapping for manage-asset to now get the assets.title,
assets.id, and type.name to be used for queries and comparisons.

Made a method to get the jsonb object containing the attribute data for
the asset as a doubly array where each item contains the pair of asset
type and data.