Feat: adapted the submitAsset function into an updateAsset function.
The new updateAsset function now queries the database to update an asset
where it's id matches the id given to the function.

There is a branch which checks if a paramter corresponding to the name
of the asset is given in the json object; the correct sql query is
triggered based on this.