Feat: adjusted validation to ignore checks regarding renaming.
The validation for labels and their corresponding entities now ignore
entities corresponding to renaming an asset.

If the checkbox to rename an asset is not checked, then the validation
ignores the rename field and checkbox values.

This is to allow easier manipulation of the jsonNode object in the java
code.